fixes #1216 fixes #1255 various fixes and improvements
