+++ library/modules/_bigip_traffic_group.py
@@ -1,0 +1,1 @@
+bigip_device_traffic_group.py
