--- library/modules/bigip_traffic_group.py
@@ -1,407 +1,0 @@
-#!/usr/bin/python
-# -*- coding: utf-8 -*-
-
-# Copyright: (c) 2017, F5 Networks Inc.
-# GNU General Public License v3.0 (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
-
-from __future__ import absolute_import, division, print_function
-__metaclass__ = type
-
-
-ANSIBLE_METADATA = {'metadata_version': '1.1',
-                    'status': ['preview'],
-                    'supported_by': 'certified'}
-
-DOCUMENTATION = r'''
----
-module: bigip_traffic_group
-short_description: Manages traffic groups on BIG-IP
-description:
-  - Supports managing traffic groups and their attributes on a BIG-IP.
-version_added: 2.5
-options:
-  name:
-    description:
-      - The name of the traffic group.
-    type: str
-    required: True
-  partition:
-    description:
-      - Device partition to manage resources on.
-    type: str
-    default: Common
-  state:
-    description:
-      - When C(present), ensures that the traffic group exists.
-      - When C(absent), ensures the traffic group is removed.
-    type: str
-    choices:
-      - present
-      - absent
-    default: present
-  mac_address:
-    description:
-      - Specifies the floating Media Access Control (MAC) address associated with the floating IP addresses
-        defined for a traffic group.
-      - Primarily, a MAC masquerade address minimizes ARP communications or dropped packets as a result of failover.
-      - A MAC masquerade address ensures that any traffic destined for a specific traffic group reaches an available
-        device after failover, which happens because along with the traffic group, the MAC masquerade address floats
-        to the available device.
-      - Without a MAC masquerade address, the sending host must learn the MAC address for a newly-active device,
-        either by sending an ARP request or by relying on the gratuitous ARP from the newly-active device.
-      - To unset the MAC address, specify an empty value (C("")) to this parameter.
-    type: str
-    version_added: 2.6
-extends_documentation_fragment: f5
-author:
-  - Tim Rupp (@caphrim007)
-  - Wojciech Wypior (@wojtek0806)
-'''
-
-EXAMPLES = r'''
-- name: Create a traffic group
-  bigip_traffic_group:
-    name: foo
-    state: present
-    provider:
-      user: admin
-      password: secret
-      server: lb.mydomain.com
-  delegate_to: localhost
-'''
-
-RETURN = r'''
-# only common fields returned
-'''
-
-from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.basic import env_fallback
-
-try:
-    from library.module_utils.network.f5.bigip import F5RestClient
-    from library.module_utils.network.f5.common import F5ModuleError
-    from library.module_utils.network.f5.common import AnsibleF5Parameters
-    from library.module_utils.network.f5.common import cleanup_tokens
-    from library.module_utils.network.f5.common import f5_argument_spec
-    from library.module_utils.network.f5.common import transform_name
-    from library.module_utils.network.f5.common import exit_json
-    from library.module_utils.network.f5.common import fail_json
-except ImportError:
-    from ansible.module_utils.network.f5.bigip import F5RestClient
-    from ansible.module_utils.network.f5.common import F5ModuleError
-    from ansible.module_utils.network.f5.common import AnsibleF5Parameters
-    from ansible.module_utils.network.f5.common import cleanup_tokens
-    from ansible.module_utils.network.f5.common import f5_argument_spec
-    from ansible.module_utils.network.f5.common import transform_name
-    from ansible.module_utils.network.f5.common import exit_json
-    from ansible.module_utils.network.f5.common import fail_json
-
-
-class Parameters(AnsibleF5Parameters):
-    api_map = {
-        'mac': 'mac_address',
-    }
-
-    api_attributes = [
-        'mac',
-    ]
-
-    returnables = [
-        'mac_address',
-    ]
-
-    updatables = [
-        'mac_address',
-    ]
-
-
-class ApiParameters(Parameters):
-    pass
-
-
-class ModuleParameters(Parameters):
-    @property
-    def mac_address(self):
-        if self._values['mac_address'] is None:
-            return None
-        if self._values['mac_address'] == '':
-            return 'none'
-        return self._values['mac_address']
-
-
-class Changes(Parameters):
-    def to_return(self):
-        result = {}
-        for returnable in self.returnables:
-            result[returnable] = getattr(self, returnable)
-        result = self._filter_params(result)
-        return result
-
-
-class UsableChanges(Changes):
-    pass
-
-
-class ReportableChanges(Changes):
-    pass
-
-
-class Difference(object):
-    def __init__(self, want, have=None):
-        self.want = want
-        self.have = have
-
-    def compare(self, param):
-        try:
-            result = getattr(self, param)
-            return result
-        except AttributeError:
-            return self.__default(param)
-
-    def __default(self, param):
-        attr1 = getattr(self.want, param)
-        try:
-            attr2 = getattr(self.have, param)
-            if attr1 != attr2:
-                return attr1
-        except AttributeError:
-            return attr1
-
-    @property
-    def partition(self):
-        raise F5ModuleError(
-            "Partition cannot be changed for a traffic group. Only /Common is allowed."
-        )
-
-
-class ModuleManager(object):
-    def __init__(self, *args, **kwargs):
-        self.module = kwargs.get('module', None)
-        self.client = kwargs.get('client', None)
-        self.have = None
-        self.want = ModuleParameters(params=self.module.params)
-        self.changes = UsableChanges()
-
-    def _set_changed_options(self):
-        changed = {}
-        for key in Parameters.returnables:
-            if getattr(self.want, key) is not None:
-                changed[key] = getattr(self.want, key)
-        if changed:
-            self.changes = UsableChanges(params=changed)
-
-    def _update_changed_options(self):
-        diff = Difference(self.want, self.have)
-        updatables = Parameters.updatables
-        changed = dict()
-        for k in updatables:
-            change = diff.compare(k)
-            if change is None:
-                continue
-            else:
-                if isinstance(change, dict):
-                    changed.update(change)
-                else:
-                    changed[k] = change
-        if changed:
-            self.changes = UsableChanges(params=changed)
-            return True
-        return False
-
-    def _announce_deprecations(self, result):
-        warnings = result.pop('__warnings', [])
-        for warning in warnings:
-            self.module.deprecate(
-                msg=warning['msg'],
-                version=warning['version']
-            )
-
-    def exec_module(self):
-        changed = False
-        result = dict()
-        state = self.want.state
-
-        if state == "present":
-            changed = self.present()
-        elif state == "absent":
-            changed = self.absent()
-
-        reportable = ReportableChanges(params=self.changes.to_return())
-        changes = reportable.to_return()
-        result.update(**changes)
-        result.update(dict(changed=changed))
-        self._announce_deprecations(result)
-        return result
-
-    def present(self):
-        if self.exists():
-            return self.update()
-        else:
-            return self.create()
-
-    def absent(self):
-        if self.exists():
-            return self.remove()
-        return False
-
-    def should_update(self):
-        result = self._update_changed_options()
-        if result:
-            return True
-        return False
-
-    def update(self):
-        self.have = self.read_current_from_device()
-        if not self.should_update():
-            return False
-        if self.module.check_mode:
-            return True
-        self.update_on_device()
-        return True
-
-    def remove(self):
-        if self.module.check_mode:
-            return True
-        self.remove_from_device()
-        if self.exists():
-            raise F5ModuleError("Failed to delete the resource.")
-        return True
-
-    def create(self):
-        self._set_changed_options()
-        if self.want.partition.lower().strip('/') != 'common':
-            raise F5ModuleError(
-                "Traffic groups can only be created in the /Common partition"
-            )
-        if self.module.check_mode:
-            return True
-        self.create_on_device()
-        return True
-
-    def exists(self):
-        uri = "https://{0}:{1}/mgmt/tm/cm/traffic-group/{2}".format(
-            self.client.provider['server'],
-            self.client.provider['server_port'],
-            transform_name(self.want.partition, self.want.name)
-        )
-        resp = self.client.api.get(uri)
-        try:
-            response = resp.json()
-        except ValueError:
-            return False
-        if resp.status == 404 or 'code' in response and response['code'] == 404:
-            return False
-        return True
-
-    def create_on_device(self):
-        params = self.changes.api_params()
-        params['name'] = self.want.name
-        params['partition'] = self.want.partition
-        uri = "https://{0}:{1}/mgmt/tm/cm/traffic-group/".format(
-            self.client.provider['server'],
-            self.client.provider['server_port'],
-        )
-        resp = self.client.api.post(uri, json=params)
-        try:
-            response = resp.json()
-        except ValueError as ex:
-            raise F5ModuleError(str(ex))
-
-        if 'code' in response and response['code'] in [400, 403]:
-            if 'message' in response:
-                raise F5ModuleError(response['message'])
-            else:
-                raise F5ModuleError(resp.content)
-        return response['selfLink']
-
-    def update_on_device(self):
-        params = self.changes.api_params()
-        uri = "https://{0}:{1}/mgmt/tm/cm/traffic-group/{2}".format(
-            self.client.provider['server'],
-            self.client.provider['server_port'],
-            transform_name(self.want.partition, self.want.name)
-        )
-        resp = self.client.api.patch(uri, json=params)
-        try:
-            response = resp.json()
-        except ValueError as ex:
-            raise F5ModuleError(str(ex))
-
-        if 'code' in response and response['code'] == 400:
-            if 'message' in response:
-                raise F5ModuleError(response['message'])
-            else:
-                raise F5ModuleError(resp.content)
-
-    def read_current_from_device(self):
-        uri = "https://{0}:{1}/mgmt/tm/cm/traffic-group/{2}".format(
-            self.client.provider['server'],
-            self.client.provider['server_port'],
-            transform_name(self.want.partition, self.want.name)
-        )
-
-        resp = self.client.api.get(uri)
-
-        try:
-            response = resp.json()
-        except ValueError as ex:
-            raise F5ModuleError(str(ex))
-
-        if 'code' in response and response['code'] == 400:
-            if 'message' in response:
-                raise F5ModuleError(response['message'])
-            else:
-                raise F5ModuleError(resp.content)
-        return ApiParameters(params=response)
-
-    def remove_from_device(self):
-        uri = "https://{0}:{1}/mgmt/tm/cm/traffic-group/{2}".format(
-            self.client.provider['server'],
-            self.client.provider['server_port'],
-            transform_name(self.want.partition, self.want.name)
-        )
-        response = self.client.api.delete(uri)
-        if response.status == 200:
-            return True
-        raise F5ModuleError(response.content)
-
-
-class ArgumentSpec(object):
-    def __init__(self):
-        self.supports_check_mode = True
-        argument_spec = dict(
-            name=dict(required=True),
-            state=dict(default='present', choices=['absent', 'present']),
-            partition=dict(
-                default='Common',
-                fallback=(env_fallback, ['F5_PARTITION'])
-            ),
-            mac_address=dict()
-        )
-        self.argument_spec = {}
-        self.argument_spec.update(f5_argument_spec)
-        self.argument_spec.update(argument_spec)
-
-
-def main():
-    spec = ArgumentSpec()
-
-    module = AnsibleModule(
-        argument_spec=spec.argument_spec,
-        supports_check_mode=spec.supports_check_mode
-    )
-
-    client = F5RestClient(**module.params)
-
-    try:
-        mm = ModuleManager(module=module, client=client)
-        results = mm.exec_module()
-        cleanup_tokens(client)
-        exit_json(module, results, client)
-    except F5ModuleError as ex:
-        cleanup_tokens(client)
-        fail_json(module, ex, client)
-
-
-if __name__ == '__main__':
-    main()
