+++ test/integration/bigip_device_ha_group.yaml
@@ -26,7 +26,7 @@
 
   vars:
     __metadata__:
-      upstream: no
+      upstream: yes
       version: 1.2
       tested_versions:
         - NA
