+++ test/integration/bigip_device_traffic_group.yaml
@@ -1,25 +1,25 @@
 ---
 
-# Test the bigip_traffic_group module
+# Test the bigip_device_traffic_group module
 #
 # Running this playbook assumes that you have a BIG-IP installation at the
 # ready to receive the commands issued in this Playbook.
 #
 # This module will run tests against a BIG-IP host to verify that the
-# bigip_traffic_group module behaves as expected.
+# bigip_device_traffic_group module behaves as expected.
 #
 # Usage:
 #
-#    ansible-playbook -i notahost, test/integration/bigip_traffic_group.yaml
+#    ansible-playbook -i notahost, test/integration/bigip_device_traffic_group.yaml
 #
 # Examples:
 #
 #    Run all tests on the {module} module
 #
-#    ansible-playbook -i notahost, test/integration/bigip_traffic_group.yaml
+#    ansible-playbook -i notahost, test/integration/bigip_device_traffic_group.yaml
 #
 
-- name: Metadata of bigip_traffic_group
+- name: Metadata of bigip_device_traffic_group
   hosts: localhost
   connection: local
   gather_facts: false
@@ -33,7 +33,7 @@
       tested_platforms:
         - NA
 
-- name: Test the bigip_traffic_group module
+- name: Test the bigip_device_traffic_group module
   hosts: "f5-test[0]"
   connection: local
   any_errors_fatal: true
@@ -49,4 +49,4 @@
     F5_VALIDATE_CERTS: "{{ validate_certs }}"
 
   roles:
-    - bigip_traffic_group
+    - bigip_device_traffic_group
